+++ cmd/importer/main.go
@@ -11,7 +11,7 @@ import (
 	"github.com/spkg/bom"
 	"github.com/uptrace/bun"
 
-	"github.com/shiroemons/touhou_arrangement_chronicle/internal/infra/store"
+	"github.com/shiroemons/touhou_arrangement_chronicle/internal/infrastructure/store"
 )
 
 type Importer struct {
